+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- Handle multiple resources for same url [#49](https://github.com/opendatateam/udata-piwik/pull/49)
 
 ## 1.1.0 (2018-03-13)